pyproject.toml
[tool.poetry]
name = "Snowflake_Build_Artifacts"
version = "0.1.0"
description = ""
authors = ["sebastian hansen <sebastian.hansen@dataalchemy.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
pyarrow = "6.0.0"
snowflake-connector-python = "^2.8.0"
pandas = "^1.5.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
